fix(review): split oversized a11y modules and cover xs target size
Extract the dev VITE_GITHUB_PAT auto-login into auth/devAutoLogin.ts (useGithubAuth drops back under 1000 lines) and split AccessibilityPage into pages/accessibility/{Vpat,Contrast,Statement}Section + reports/PrintableReport, with a shared data module that reuses the canonical ContrastAuditJson/VpatReportJson producer types instead of a local re-declaration. Add the shipped xs Button size (incl. an icon-only xs circle) to the 2.5.8 target-size guard so the VPAT Supports claim covers every action size.
