+++ web/src/pages/accessibility/ContrastSection.tsx
@@ -1,0 +1,350 @@
+import { useEffect, useRef, useState } from "react"
+import { useTranslation } from "react-i18next"
+import { Info } from "lucide-react"
+
+import { Alert, Badge, Card, Modal } from "@/components/ui"
+
+import {
+  STATUS_TONE,
+  tabClass,
+  useContrastAudit,
+  type Row,
+  type AuditTheme,
+} from "./data"
+
+// A live preview of the pair — the actual foreground text on the actual
+// surface. Rendered as a button that opens the detail modal.
+function ColorPreviewButton({ row, onOpen }: { row: Row; onOpen: () => void }) {
+  const { t } = useTranslation()
+  const isText = row.kind === "text"
+  return (
+    <button
+      type="button"
+      onClick={onOpen}
+      title={t("accessibility.preview.open")}
+      aria-label={t("accessibility.preview.openFor", { pair: row.id })}
+      className="flex h-9 w-14 cursor-pointer items-center justify-center rounded border border-base-300 text-sm font-semibold transition hover:ring-2 hover:ring-primary/40"
+      style={{ backgroundColor: row.bgHex, color: row.fgHex }}
+    >
+      <span aria-hidden="true">{isText ? "Aa" : "▭"}</span>
+    </button>
+  )
+}
+
+// One reference chip: a small swatch + role label + hex. Reference only — the
+// combined sample above is what actually shows the contrast.
+function ColorRef({ hex, role }: { hex: string; role: string }) {
+  return (
+    <div className="flex items-center gap-2">
+      <span
+        className="size-8 shrink-0 rounded border border-base-300"
+        style={{ backgroundColor: hex }}
+        aria-hidden="true"
+      />
+      <div className="flex flex-col">
+        <span className="text-xs uppercase tracking-wide text-base-content/60">
+          {role}
+        </span>
+        <span className="font-mono text-sm">{hex}</span>
+      </div>
+    </div>
+  )
+}
+
+function StatusCell({ row }: { row: Row }) {
+  const { t } = useTranslation()
+  return (
+    <div className="flex items-center gap-1">
+      <Badge tone={STATUS_TONE[row.status]}>
+        {t(`accessibility.status.${row.status}`)}
+      </Badge>
+      {row.withinMargin && (
+        <span
+          className="tooltip tooltip-left cursor-help text-base-content/60 hover:text-base-content"
+          data-tip={t("accessibility.marginInfo", {
+            ratio: row.ratio.toFixed(2),
+            floor: row.floor,
+            margin: row.margin,
+          })}
+        >
+          <Info aria-hidden="true" className="size-4" />
+        </span>
+      )}
+    </div>
+  )
+}
+
+function PairDetailModal({ row, onClose }: { row: Row; onClose: () => void }) {
+  const { t } = useTranslation()
+  const dialogRef = useRef<HTMLDialogElement | null>(null)
+
+  // Mounted only while a row is selected (caller gates + remounts via `key`), so
+  // open once imperatively. This avoids the controlled-`open` + conditional-
+  // content race that emptied the box mid close-animation.
+  useEffect(() => {
+    dialogRef.current?.showModal()
+  }, [])
+
+  return (
+    <Modal dialogRef={dialogRef} onClose={onClose} size="lg">
+      <div className="flex flex-col gap-4">
+        <div>
+          <h3 className="font-mono text-sm font-semibold">{row.id}</h3>
+          <p className="text-sm text-base-content/70">{row.label}</p>
+        </div>
+
+        {/* The combined sample: real foreground on the real surface — the way
+            WebAIM and other contrast tools preview a pair, because contrast is
+            a property of the combination, not two isolated swatches. */}
+        <div
+          className="flex flex-col items-center justify-center gap-2 rounded-box border border-base-300 p-8"
+          style={{ backgroundColor: row.bgHex, color: row.fgHex }}
+        >
+          <span className="text-2xl font-semibold" aria-hidden="true">
+            {t("accessibility.detail.sampleHeading")}
+          </span>
+          <span className="text-sm" aria-hidden="true">
+            {t("accessibility.detail.sampleBody")}
+          </span>
+        </div>
+
+        <div className="flex flex-col gap-2 sm:flex-row sm:justify-between">
+          <ColorRef
+            hex={row.fgHex}
+            role={t("accessibility.detail.foreground")}
+          />
+          <ColorRef
+            hex={row.bgHex}
+            role={t("accessibility.detail.background")}
+          />
+        </div>
+
+        <div className="rounded-box border border-base-300 p-4">
+          <div className="flex items-baseline justify-between gap-3">
+            <span className="text-sm text-base-content/70">
+              {t("accessibility.detail.ratio")}
+            </span>
+            <span className="font-mono text-2xl font-bold tabular-nums">
+              {row.ratio.toFixed(2)}:1
+            </span>
+          </div>
+          <div className="mt-2 flex items-center justify-between gap-3">
+            <span className="text-sm text-base-content/70">
+              {t("accessibility.detail.minimum", { floor: row.floor })}
+            </span>
+            <StatusCell row={row} />
+          </div>
+          {row.withinMargin && (
+            <p className="mt-2 text-sm text-base-content/70">
+              {t("accessibility.marginInfo", {
+                ratio: row.ratio.toFixed(2),
+                floor: row.floor,
+                margin: row.margin,
+              })}
+            </p>
+          )}
+        </div>
+      </div>
+    </Modal>
+  )
+}
+
+function ThemeTable({
+  theme,
+  onOpenRow,
+}: {
+  theme: AuditTheme
+  onOpenRow: (row: Row) => void
+}) {
+  const { t } = useTranslation()
+  return (
+    <Card radius="xl" shadow={false}>
+      <Card.Body className="gap-3 p-4">
+        <div className="overflow-x-auto">
+          <table className="table table-sm">
+            <thead>
+              <tr>
+                <th>{t("accessibility.col.preview")}</th>
+                <th>{t("accessibility.col.pair")}</th>
+                <th>{t("accessibility.col.description")}</th>
+                <th className="text-end">{t("accessibility.col.ratio")}</th>
+                <th className="text-end">{t("accessibility.col.floor")}</th>
+                <th>{t("accessibility.col.status")}</th>
+              </tr>
+            </thead>
+            <tbody>
+              {theme.rows.map((r) => (
+                <tr key={r.id}>
+                  <td>
+                    <ColorPreviewButton row={r} onOpen={() => onOpenRow(r)} />
+                  </td>
+                  <td className="font-mono text-xs">{r.id}</td>
+                  <td>
+                    {r.label}
+                    <span className="block text-xs text-base-content/60">
+                      {r.size}
+                    </span>
+                  </td>
+                  <td className="text-end font-mono tabular-nums">
+                    {r.ratio.toFixed(2)}:1
+                  </td>
+                  <td className="text-end font-mono tabular-nums text-base-content/70">
+                    {r.floor}:1
+                  </td>
+                  <td>
+                    <StatusCell row={r} />
+                  </td>
+                </tr>
+              ))}
+            </tbody>
+          </table>
+        </div>
+      </Card.Body>
+    </Card>
+  )
+}
+
+export function ContrastSection() {
+  const { t } = useTranslation()
+  const { data: audit = null, isError: error } = useContrastAudit()
+  // null = "no explicit pick yet"; the shown theme then defaults to the first.
+  // Deriving the shown theme (rather than syncing it in an effect) keeps the
+  // user's later choice sticky without a setState-in-effect.
+  const [pickedTheme, setPickedTheme] = useState<string | null>(null)
+  const [selectedRow, setSelectedRow] = useState<Row | null>(null)
+
+  const activeTheme = pickedTheme ?? audit?.themes[0]?.theme ?? null
+  const shownTheme = audit?.themes.find((th) => th.theme === activeTheme)
+  const marginCount = audit?.summary.marginMisses ?? 0
+
+  return (
+    <section className="flex flex-col gap-4" aria-labelledby="contrast-heading">
+      <h2 id="contrast-heading" className="sr-only">
+        {t("accessibility.title")}
+      </h2>
+
+      {error && <Alert tone="error">{t("accessibility.loadError")}</Alert>}
+
+      {!error && !audit && (
+        <div className="skeleton skeleton-shimmer h-40 w-full rounded-box" />
+      )}
+
+      {audit && (
+        <>
+          <p className="text-sm text-base-content/70">
+            {!audit.summary.allPass
+              ? t("accessibility.summaryFail", {
+                  failures: audit.summary.failures,
+                })
+              : marginCount > 0
+                ? t("accessibility.summaryPassMargin", {
+                    total: audit.summary.total,
+                    count: marginCount,
+                  })
+                : t("accessibility.summaryPass", {
+                    total: audit.summary.total,
+                  })}
+          </p>
+
+          <div className="flex flex-wrap items-center gap-3">
+            <div role="tablist" className="tabs-boxed tabs w-fit">
+              {audit.themes.map((th) => (
+                <button
+                  key={th.theme}
+                  type="button"
+                  role="tab"
+                  aria-selected={th.theme === activeTheme}
+                  className={tabClass(th.theme === activeTheme)}
+                  onClick={() => setPickedTheme(th.theme)}
+                >
+                  {th.label}
+                </button>
+              ))}
+            </div>
+          </div>
+
+          {shownTheme && (
+            <ThemeTable theme={shownTheme} onOpenRow={setSelectedRow} />
+          )}
+
+          <details className="collapse-arrow collapse rounded-box border border-base-300 bg-base-100">
+            <summary className="collapse-title text-sm font-medium">
+              {t("accessibility.criteria.aboutHeading")}
+            </summary>
+            <div className="collapse-content flex flex-col gap-4 text-sm text-base-content/70">
+              <p>{t("accessibility.criteria.standardText")}</p>
+              <div>
+                <p className="font-medium text-base-content">
+                  {t("accessibility.criteria.thresholdsHeading")}
+                </p>
+                <ul className="mt-1 list-disc ps-5">
+                  <li>
+                    {t("accessibility.criteria.bodyText", {
+                      body: audit.thresholds.body,
+                    })}
+                  </li>
+                  <li>
+                    {t("accessibility.criteria.largeText", {
+                      large: audit.thresholds.large,
+                    })}
+                  </li>
+                  <li>
+                    {t("accessibility.criteria.nonText", {
+                      nonText: audit.thresholds.nonText,
+                    })}
+                  </li>
+                </ul>
+              </div>
+              <div>
+                <p className="font-medium text-base-content">
+                  {t("accessibility.criteria.marginHeading")}
+                </p>
+                <p className="mt-1">{t("accessibility.criteria.marginWhy")}</p>
+              </div>
+              <div>
+                <p className="font-medium text-base-content">
+                  {t("accessibility.legend.heading")}
+                </p>
+                <div className="mt-1 flex flex-col gap-1">
+                  <span className="flex items-center gap-2">
+                    <Badge tone="success">
+                      {t("accessibility.status.pass")}
+                    </Badge>
+                    {t("accessibility.legend.pass")}
+                  </span>
+                  <span className="flex items-center gap-2">
+                    <Info aria-hidden="true" className="size-4" />
+                    {t("accessibility.legend.passMargin")}
+                  </span>
+                  <span className="flex items-center gap-2">
+                    <Badge tone="error">{t("accessibility.status.fail")}</Badge>
+                    {t("accessibility.legend.fail")}
+                  </span>
+                  <span className="flex items-center gap-2">
+                    <Badge tone="neutral">
+                      {t("accessibility.status.exempt")}
+                    </Badge>
+                    {t("accessibility.legend.exempt")}
+                  </span>
+                </div>
+              </div>
+              <p className="text-xs">
+                {t("accessibility.criteria.generated", {
+                  generated: audit.generated,
+                })}
+              </p>
+            </div>
+          </details>
+        </>
+      )}
+
+      {selectedRow && (
+        <PairDetailModal
+          key={selectedRow.id}
+          row={selectedRow}
+          onClose={() => setSelectedRow(null)}
+        />
+      )}
+    </section>
+  )
+}
