+++ web/src/test/targetSize.browser.test.tsx
@@ -11,8 +11,11 @@ import { rect, renderInViewport, setupBrowserA11y } from "./browserA11y"
 setupBrowserA11y()
 
 describe("2.5.8 Target Size — shared Button primitive", () => {
-  // md and sm are the default action sizes across the app; both must clear 24px.
-  it.each(["md", "sm"] as const)(
+  // xs/sm/md are the action sizes the app ships; every one must clear 24px, so
+  // measure them all (not just the defaults) — the VPAT 2.5.8 "Supports" claim
+  // covers the primitive's action sizes, so an unmeasured shipped size would
+  // let the claim outrun the check.
+  it.each(["md", "sm", "xs"] as const)(
     "a %s Button meets the 24x24 minimum",
     (size) => {
       const { getByRole } = renderInViewport(<Button size={size}>Save</Button>)
@@ -23,6 +26,18 @@ describe("2.5.8 Target Size — shared Button primitive", () => {
       ).toBe(true)
     },
   )
+
+  // The smallest icon-only target the app renders: an xs circle. Icon-only + the
+  // smallest size is the worst case for the 24x24 floor, so pin it explicitly.
+  it("an icon-only xs circle Button meets the 24x24 minimum", () => {
+    const { getByRole } = renderInViewport(
+      <Button shape="circle" size="xs" aria-label="Remove">
+        <span aria-hidden="true" className="size-3" />
+      </Button>,
+    )
+    const { width, height } = rect(getByRole("button"))
+    expect(meetsTargetSize(width, height), `${width}x${height}`).toBe(true)
+  })
 
   it("an icon-only circle Button meets the 24x24 minimum", () => {
     const { getByRole } = renderInViewport(
